ai_news_agent/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.2
feedparser>=6.0.10
python-dotenv>=1.0.0
openai>=1.3.7
schedule>=1.2.0
aiohttp>=3.9.1
jinja2>=3.1.2
python-multipart>=0.0.6
pydantic>=2.8.0